Move ExitDialog to its own submodule
